smart-building-rag/services/api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
psycopg[binary]==3.2.1
requests==2.32.3
chromadb==0.5.3
sentence-transformers==3.0.1
rank-bm25==0.2.2
numpy==1.26.4
pymupdf==1.24.7
unstructured==0.14.6
tqdm==4.66.4
python-dotenv==1.0.1
# Optional: OpenAI / Ollama client
openai==1.35.10
httpx==0.27.0
